backend/requirements.txt
# === Core Framework ===
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# === OpenAI & LLM ===
openai==1.3.5

# === Document Processing ===
pdfplumber==0.10.3
PyPDF2==3.0.1

# === Vector Store & FAISS ===
faiss-cpu>=1.8.0
numpy>=1.26.0

# === Audio Processing ===
librosa==0.10.0
pydub==0.25.1

# === HTTP Requests ===
requests==2.31.0

# === Data Validation ===
pydantic==2.4.2
pydantic-settings==2.1.0

# === Utilities ===
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0

# === Development ===
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1
